libs/k3sapp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "k3sapp"
version = "0.1.0"
description = "K3s Platform CLI for traditional Dockerfile-based applications"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "K3s Platform" }]
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.20",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
k3sapp = "k3sapp.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["k3sapp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
